requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
gunicorn==21.2.0
pydantic==2.12.5
pyswisseph==2.10.3.2
pytz==2025.2
python-dateutil==2.9.0.post0
requests==2.32.5

